Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
pydantic
pydantic-settings
groq
email-validator